+++ aws-auth-cognito/src/main/java/com/amplifyframework/auth/cognito/AWSCognitoAuthSession.kt
@@ -20,10 +20,14 @@ import com.amplifyframework.auth.AWSCognitoUserPoolTokens
 import com.amplifyframework.auth.AWSCredentials
 import com.amplifyframework.auth.AuthException
 import com.amplifyframework.auth.AuthSession
+import com.amplifyframework.auth.CognitoCredentials
+import com.amplifyframework.auth.CognitoUserPoolTokens
+import com.amplifyframework.auth.cognito.helpers.JWTParser
 import com.amplifyframework.auth.cognito.helpers.SessionHelper
 import com.amplifyframework.auth.exceptions.ConfigurationException
 import com.amplifyframework.auth.exceptions.InvalidStateException
 import com.amplifyframework.auth.exceptions.SignedOutException
+import com.amplifyframework.auth.exceptions.UnknownException
 import com.amplifyframework.auth.result.AuthSessionResult
 import com.amplifyframework.statemachine.codegen.data.AmplifyCredential
 
@@ -58,6 +62,38 @@ internal fun AmplifyCredential.isValid(): Boolean {
 internal fun AmplifyCredential.getCognitoSession(
     exception: AuthException = SignedOutException()
 ): AWSCognitoAuthSession {
+
+    fun getCredentialsResult(awsCredentials: CognitoCredentials): AuthSessionResult<AWSCredentials> =
+        with(awsCredentials) {
+            AWSCredentials.createAWSCredentials(accessKeyId, secretAccessKey, sessionToken, expiration)
+        }?.let {
+            AuthSessionResult.success(it)
+        } ?: AuthSessionResult.failure(UnknownException("Failed to fetch AWS credentials."))
+
+    fun getIdentityIdResult(identityId: String): AuthSessionResult<String> {
+        return if (identityId.isNotEmpty()) AuthSessionResult.success(identityId)
+        else AuthSessionResult.failure(UnknownException("Failed to fetch identity id."))
+    }
+
+    fun getUserSubResult(userPoolTokens: CognitoUserPoolTokens?): AuthSessionResult<String> {
+        return try {
+            AuthSessionResult.success(userPoolTokens?.accessToken?.let { JWTParser.getClaim(it, "sub") })
+        } catch (e: Exception) {
+            AuthSessionResult.failure(UnknownException(cause = e))
+        }
+    }
+
+    fun getUserPoolTokensResult(cognitoUserPoolTokens: CognitoUserPoolTokens):
+        AuthSessionResult<AWSCognitoUserPoolTokens> {
+        return AuthSessionResult.success(
+            AWSCognitoUserPoolTokens(
+                accessToken = cognitoUserPoolTokens.accessToken,
+                idToken = cognitoUserPoolTokens.idToken,
+                refreshToken = cognitoUserPoolTokens.refreshToken
+            )
+        )
+    }
+
     return when (this) {
         is AmplifyCredential.UserPool -> AWSCognitoAuthSession(
             true,
@@ -73,20 +109,20 @@ internal fun AmplifyCredential.getCognitoSession(
                     "Cognito Identity not configured. Please check amplifyconfiguration.json file."
                 )
             ),
-            userSubResult = AWSCognitoAuthSession.getUserSubResult(signedInData.cognitoUserPoolTokens),
-            userPoolTokensResult = AWSCognitoAuthSession.getUserPoolTokensResult(signedInData.cognitoUserPoolTokens)
+            userSubResult = getUserSubResult(signedInData.cognitoUserPoolTokens),
+            userPoolTokensResult = getUserPoolTokensResult(signedInData.cognitoUserPoolTokens)
         )
         is AmplifyCredential.UserAndIdentityPool -> AWSCognitoAuthSession(
             true,
-            identityIdResult = AWSCognitoAuthSession.getIdentityIdResult(identityId),
-            awsCredentialsResult = AWSCognitoAuthSession.getCredentialsResult(credentials),
-            userSubResult = AWSCognitoAuthSession.getUserSubResult(signedInData.cognitoUserPoolTokens),
-            userPoolTokensResult = AWSCognitoAuthSession.getUserPoolTokensResult(signedInData.cognitoUserPoolTokens)
+            identityIdResult = getIdentityIdResult(identityId),
+            awsCredentialsResult = getCredentialsResult(credentials),
+            userSubResult = getUserSubResult(signedInData.cognitoUserPoolTokens),
+            userPoolTokensResult = getUserPoolTokensResult(signedInData.cognitoUserPoolTokens)
         )
         is AmplifyCredential.IdentityPool -> AWSCognitoAuthSession(
             false,
-            identityIdResult = AWSCognitoAuthSession.getIdentityIdResult(identityId),
-            awsCredentialsResult = AWSCognitoAuthSession.getCredentialsResult(credentials),
+            identityIdResult = getIdentityIdResult(identityId),
+            awsCredentialsResult = getCredentialsResult(credentials),
             userSubResult = AuthSessionResult.failure(SignedOutException()),
             userPoolTokensResult = AuthSessionResult.failure(SignedOutException())
         )
@@ -97,8 +133,8 @@ internal fun AmplifyCredential.getCognitoSession(
             )
             AWSCognitoAuthSession(
                 true,
-                identityIdResult = AWSCognitoAuthSession.getIdentityIdResult(identityId),
-                awsCredentialsResult = AWSCognitoAuthSession.getCredentialsResult(credentials),
+                identityIdResult = getIdentityIdResult(identityId),
+                awsCredentialsResult = getCredentialsResult(credentials),
                 userSubResult = AuthSessionResult.failure(userPoolException),
                 userPoolTokensResult = AuthSessionResult.failure(userPoolException)
             )
